[Refactor] Move exit_coder from cmd to pkg/exit_coder
