+++ cmd/nerdctl/run.go
@@ -42,6 +42,7 @@ import (
 	"github.com/containerd/nerdctl/pkg/api/types"
 	"github.com/containerd/nerdctl/pkg/clientutil"
 	"github.com/containerd/nerdctl/pkg/defaults"
+	"github.com/containerd/nerdctl/pkg/errutil"
 	"github.com/containerd/nerdctl/pkg/idgen"
 	"github.com/containerd/nerdctl/pkg/imgutil"
 	"github.com/containerd/nerdctl/pkg/inspecttypes/dockercompat"
@@ -398,9 +399,7 @@ func runAction(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	if code != 0 {
-		return ExitCodeError{
-			exitCode: int(code),
-		}
+		return errutil.NewExitCoderErr(int(code))
 	}
 	return nil
 }
